+++ src/static/style.css
@@ -45,6 +45,14 @@ table.visuals {
 .hidden {
 	display: none;
 }
+.background {
+	background-color: rgba(0, 0, 0, 0.9);
+	bottom: 0;
+	left: 0;
+	position: fixed;
+	right: 0;
+	top: 0;
+}
 
 .buttons {
 	text-align: center;
