+++ src/templates/data-edit.html
@@ -26,14 +26,6 @@
 .matrix td.green {
 	background-color: green;
 }
-.background {
-	background-color: rgba(0, 0, 0, 0.9);
-	bottom: 0;
-	left: 0;
-	position: fixed;
-	right: 0;
-	top: 0;
-}
 .editor {
 	background-color: white;
 	border-radius: 1em;
