+++ src/templates/match.html
@@ -57,10 +57,15 @@
 			ui.txtTheirCharity.textContent = match.their_charity;
 			ui.txtTheirAmount.textContent = (match.my_currency === match.their_currency)
 				? `${match.their_amount} ${match.their_currency}`
-				: `${match.their_amount} ${match.their_currency}`
+				: `${match.their_amount} ${match.their_currency}`;
+			ui.btnApprove.disabled = !match.can_edit;
+			ui.btnDecline.disabled = !match.can_edit;
+			ui.btnApprove.title = ui.btnDecline.title = match.can_edit ? '' : 'You have already approved or declined this match';
 		});
 
 	ui.btnApprove.onclick = () => {
+		ui.btnApprove.disabled = true;
+		ui.btnDecline.disabled = true;
 		ajax('/ajax/approve_match', { secret })
 			.then(() => {
 				//xxx say "you'll get an email soon."
@@ -69,6 +74,9 @@
 	};
 
 	ui.btnDecline.onclick = () => {
+		ui.btnApprove.disabled = true;
+		ui.btnDecline.disabled = true;
+
 		//xxx un-hide feedback form and actual decline button
 		const feedback = 'xxx not implemented yet';
 		ajax('/ajax/decline_match', { secret, feedback })
