+++ src/templates/unmatched-offers.html
@@ -41,6 +41,12 @@ summary {
 	margin-left: 1em;
 	text-decoration: underline;
 }
+#waitText {
+	color: white;
+	font-size: 150%;
+	margin-top: 5em;
+	text-align: center;
+}
 </style>
 <title>Unmatched Offers</title>
 </head>
@@ -52,6 +58,13 @@ summary {
 	Click the resulting score number on the other offer to create a match.
 </p>
 <ul id="offers"></ul>
+<div class="background hidden" id="spinner">
+	<div id="waitText">
+		Creating match...<br>
+		Sending emails...<br>
+		This will take a few seconds...
+	</div>
+</div>
 <script>
 /* globals ajax, createNode, getElementsById */
 /* jshint esversion: 6 */
@@ -65,7 +78,7 @@ summary {
 	};
 
 	function getSummary(offer) {
-		return `${offer.country}, ${offer.charity}, $${offer.amount_NZD}`;
+		return `${offer.country}, ${offer.charity}, $${offer.min_amount_NZD}-$${offer.amount_NZD}`;
 	}
 
 	function renderDetails(details) {
@@ -101,7 +114,22 @@ summary {
 		const confirmed = window.confirm(`Please confirm that you want to create a match for the following two offers:\n\n${first}\n\n${second}`);
 
 		if (confirmed) {
-			alert('not implemented yet. sorry.'); //xxx
+			ui.spinner.classList.remove('hidden');
+			ajax('/special-secret-admin/create_match', {
+				offer_a_id: state.firstOfferId,
+				offer_b_id: offerId,
+			})
+				.then(() => {
+					alert('Match created. Emails sent.\n\nRefreshing page now...');
+					window.location.reload();
+				})
+				.catch(error => {
+					console.error(error);
+					window.alert(`Unexpected error: ${error}`);
+				})
+				.then(() => {
+					ui.spinner.classList.add('hidden');
+				});
 		}
 	}
